+++ packages/code-du-travail-frontend/cypress/integration/light/glossaire.spec.ts
@@ -11,7 +11,7 @@ describe("Glossaire", () => {
     cy.contains("Définition");
     cy.contains("Sources");
 
-    cy.contains("Retour").click();
+    cy.contains("Glossaire").click();
 
     cy.contains(
       "Les définitions de ce glossaire, disponibles en surbrillance dans les textes des réponses"