+++ packages/code-du-travail-frontend/src/api/modules/tools/service.ts
@@ -1,90 +1,12 @@
-import { Tool } from "@socialgouv/cdtn-types";
-import {
-  elasticsearchClient,
-  elasticDocumentsIndex,
-  NotFoundError,
-} from "../../utils";
-import { getTools, getAllToolsQuery } from "./queries";
-import { SearchHit } from "@elastic/elasticsearch/lib/api/types";
-
-export const getAllTools = async (): Promise<Tool[]> => {
-  const body: any = getAllToolsQuery();
-  const response = await elasticsearchClient.search<any>({
-    body,
-    index: elasticDocumentsIndex,
-  });
-  if (response.hits.hits.length === 0) {
-    throw new NotFoundError({
-      message: `There is no tools that match query`,
-      name: "TOOLS_NOT_FOUND",
-      cause: null,
-    });
-  }
-  return response.hits.hits
-    .map(({ _id, _source }) => ({ ..._source, _id }))
-    .filter((tool) => tool.displayTool);
-};
-
-export const getToolsByIdsAndSlugs = async (
-  ids?: string[],
-  slugs?: string[]
-): Promise<SearchHit<Tool>[]> => {
-  const body: any = getTools(ids, slugs);
-  const response = await elasticsearchClient.search<Tool>({
-    body,
-    index: elasticDocumentsIndex,
-  });
-  if (response.hits.hits.length === 0) {
-    throw new NotFoundError({
-      message: `There is no tools that match query`,
-      name: "TOOLS_NOT_FOUND",
-      cause: null,
-    });
-  }
-  return response.hits.hits;
-};
-
-export const getToolsByIds = async (
-  cdtnIds: string[]
-): Promise<SearchHit<Tool>[]> => {
-  const body: any = getTools(undefined, undefined, cdtnIds);
-  const response = await elasticsearchClient.search<Tool>({
-    body,
-    index: elasticDocumentsIndex,
-  });
-  if (response.hits.hits.length === 0) {
-    throw new NotFoundError({
-      message: `There is no tools that match query`,
-      name: "TOOLS_NOT_FOUND",
-      cause: null,
-    });
-  }
-  return response.hits.hits;
-};
-
-export const getToolsBySlugs = async (
-  slugs: string[]
-): Promise<SearchHit<Tool>[]> => {
-  const body: any = getTools(undefined, slugs);
-  const response = await elasticsearchClient.search<Tool>({
-    body,
-    index: elasticDocumentsIndex,
-  });
-  if (response.hits.hits.length === 0) {
-    throw new NotFoundError({
-      message: `There is no tools that match query`,
-      name: "TOOLS_NOT_FOUND",
-      cause: null,
-    });
-  }
-  return response.hits.hits;
-};
+import { ElasticTool } from "@socialgouv/cdtn-types";
+import { elasticDocumentsIndex, elasticsearchClient } from "../../utils";
+import { getTools } from "./queries";
 
 export const getBySlugTools = async (
   slug: string
-): Promise<Tool | undefined> => {
+): Promise<ElasticTool | undefined> => {
   const body: any = getTools(undefined, [slug]);
-  const response = await elasticsearchClient.search<Tool>({
+  const response = await elasticsearchClient.search<ElasticTool>({
     body,
     index: elasticDocumentsIndex,
   });