+++ packages/code-du-travail-frontend/src/modules/common/Autocomplete/Autocomplete.tsx
@@ -91,10 +91,7 @@ export const Autocomplete = <K,>({
         getRootProps,
         clearSelection,
       }) => (
-        <div
-          className={`${searchContainer}`}
-          {...getRootProps({}, { suppressRefError: true })}
-        >
+        <div className={`${searchContainer}`}>
           <Input
             nativeLabelProps={getLabelProps()}
             hideLabel={isSearch}
@@ -137,6 +134,8 @@ export const Autocomplete = <K,>({
               "data-testid": dataTestId,
               placeholder,
               ref: setInputRef,
+              role: getRootProps().role,
+              "aria-expanded": getRootProps()["aria-expanded"],
               ...getInputProps(),
             }}
             className={`${fr.cx("fr-mb-0")}`}