+++ packages/code-du-travail-frontend/src/outils/service.ts
@@ -1,7 +1,9 @@
-import { Tool } from "@socialgouv/cdtn-types";
+import { ElasticTool } from "@socialgouv/cdtn-types";
 import { SITE_URL } from "../config";
 
-export const fetchTool = async (slug: string): Promise<Tool | undefined> => {
+export const fetchTool = async (
+  slug: string
+): Promise<ElasticTool | undefined> => {
   const responseContainer = await fetch(`${SITE_URL}/api/tools/${slug}`);
   if (responseContainer.ok) {
     const result = await responseContainer.json();